I want a simple way to scroll to top. Moving hand fto icon at bottom to scroll to top is awkward.
So I tried to implement similar concet as tidings apps. However I thinks the arrow need to be bigger than tiding and also centered...

Problem is I am not sure this solution is still not the best... We need testers...

First of

The icon I made is just rescaled from original jolla arrow and could be done better
Second: for some uknown reason it does not fully scroll to toå using flickered.scrollToTop() seems like a silica bug? or raise condtion sine onClicked triggered and still scrolling?

However this prof of concept that I hope can be implemented in some way cause todays behave is not good IMHO....